requirements.txt
appnope==0.1.3
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
category-encoders==2.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
cycler==0.11.0
dask==2022.7.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distributed==2022.7.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
feather-format==0.4.1
featurewiz==0.1.87
Flask==2.1.2
fonttools==4.33.3
fsspec==2022.5.0
greenlet==1.1.2
HeapDict==1.0.1
idna==3.3
imbalanced-learn==0.9.1
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
lightgbm==3.3.2
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
patsy==0.5.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
Pygments==2.12.0
pymongo==4.1.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.3.0
statsmodels==0.13.2
tblib==1.7.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
xgboost==1.6.1
xlrd==2.0.1
xmltodict==0.13.0
zict==2.2.0
zipp==3.8.0

gdown~=4.5.1
tensorflow~=2.9.1
adtk~=0.6.2
keras~=2.9.0